feat(VerifyCertificateMacOS): moves the HttpsVerifier class to separate files
